Fix bug causing connection retry infinite loop
For some adapters, if there are any problems connecting to a database,
part of the rescue process involves resetting the database connection,

Since adapters using ActiveRecord's connection pooling reset their
connection by calling the `#connect_to_new` method with the
default_database, any problems connecting to the default_database get
stuck in an infinite rescue->reset->rescue->reset... cycle.

Now, we'll only try to reset the connection if we weren't trying to
connect to the default database to begin with.