Consistently switch tenant in #switch
When the `base` argument is set to true a "base" switch is performed
which establishes a new connection on the base ActiveRecord::Base class
rather than the dynamic ConnectionPool class. However, when switching
back it did not honor this argument and so the change was not undone.

Now, the behavior is consistent within the method and a base switch will
result in a base unswitch after the block is run.