Fix broken #create method on AbstractAdapter
The create method was properly creating the databases for all tenants,
but it was not properly loading the schema. Schema loads in ActiveRecord
are run on the `ActiveRecord::Base` connection not on the connection
pool established by Apartment. Thus, the schema loads were always being
run on the primary database.

By performing a base switch, the schema load happens on the actual tenant
database. Unfortunately though, we still have to do a normal switch as
well. ActiveRecord makes a call to `ActiveRecord::SchemaMigration` in
order to create the `schema_migrations` table which will read its
connection from the configured Apartment connection, not
`ActiveRecord::Base`.

Performing both switches allows for both the schema load operation and
the `schema_migrations` table creation to happen properly.