Fix `apartment:create` rake task
This rake task was errantly dependent on the `db:migrate` task. There is
no reason to try to run migrations on a database that may not even exist
yet especially when the create task that is being invoked does a schema
load! The `environment` task is required though as the `db:migrate` task
was implicitly loading that one for us.