Fix broken Migrator tasks
This is the same issue fixed on the #create method in a previous commit:

ad2428050b976d271e3aa2bf6911ce0525668a33

In short, we need to switch the tenant on both the base
ActiveRecord::Base connection and on all other model connections. I
believe the double switch is required because of the ConnectionPool
class that we added.