Using ConnectionHandler.establish_connection
This will allow us to accomplish what we were doing in a monkey patch to
`establish_connection` in ConnectionHandling without having to do any
monkey patching. Essentially we are creating a new "Spec", in
ActiveRecord terms, by the name of the tenant and then setting the Base
class's `connection_specification_name` to the name of the tenant so
that the correct spec gets used.

This is what we were doing in the ConnectionHandling overrides, but now
we can do it directly through the ConnectionHandler.