Thread safety for ActiveRecord::ConnectionHandling
When switching tenants in a thread our current implementation would not
work properly. It would globally change the tenant for all threads which
is not thread-safe. With this implementation we should be safe in most
circumstances. The only non-thread safe operation I can foresee would be
connection removal. So a call to `remove_connection` from in a subthread
would reference the class instance variable (thread global). To my
knowledge, we are not using this method explicitly anywhere so we should
be ok.

This also maintains the original behavior for the main thread. The
behavior should only be changed in subthreads where the global value
will no longer be set, only the thread local one.

One other caveat is on a sub-subthread the tenant inherited will NOT be
the tenant from the immediately previous thread, but the one from the
main thread as the thread local variables are not inherited from the
parent thread so the global will be read. It is good practice to do an
explicit tenant switch anytime a new thread is created.